chore: 자료형 수정

+++ shop/src/main/java/com/apple/shop/Notice.java
@@ -7,7 +7,7 @@ import java.util.Date;
 @Entity
 public class Notice {
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    public Integer id;
+    public long id;
 
     @Column(length = 500)
     public String title;
